Bug 33484: Always save but restore on demand only
And remove get_columns_saved_state. I am not sure this what really
helpful, maybe in case the yml or config in the DB changed, but I don't
think we should complicate the code. If something is wrong, logout to
clear localStorage...
We will see later if we really need it.

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>