Bug 33484: Fix Vue tables
datatables.net-vue3 does not use a specific id, and does not let us pass
one. It is generated, like: DataTables_Table_0 (and increment when a new
table is created).
So it's not possible to use this id: it can change and so is not
reliable.

I have tried to pass "data-koha-table-id" with a specific id
(KohaTable_agreements). Which was working, but actually we could simply
use the table settings key to identify the tables in the URL.

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>