Bug 38116: Build patron search descriptions using DT's search settings
The goal of this patch is to move the construction of the patron search description
to datatables.js in order to make it generic and reusable.

With bug 33484 and the restoration of the state from localStorage or the
URL could lead to confusion when the table is displayed. It could also
be problematic when an hidden column has a search pattern.

Note that if a column has a dropdown list with an option selected AND is hidden,
the description won't display the value but the code (eg. ^CPL$ if
library Centerville is selected)

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>