Bug 33484: Fix libraries table
We added columns to the table but never displayed them, and they were
not part of the columns settings.
I am not sure what was the purpose of that but it looks completely
wrong. If we want to restore them it should be done correctly (ie. add
them back to the column list and add the new ones to the columns
settings).

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>