Bug 33484: Allow several tables using the same table settings - columns_settings
In some cases (at least on the suggestions page) we display several
pages using the same settings.
In that case the key "DataTables_acqui_suggestions_suggestions" is not
specific enough, we need to add the id of the table as well.

This patch does not modify koha-tmpl/intranet-tmpl/prog/js/datatables.js
/ REST API wrapper as I don't think we have occurrences that need this.

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>