Bug 33484: External filter into state - catalogue concerns
Use a checkbox to store the state. It does not make the code better but
make possible to keep and restore and state of the filter.

Not sure it's working correct on the cataloguing/concerns.pl page but
the behaviour was already not good before.
At least I think it fixes the filter/include icon in some situation.
On main I managed to get the "include resolved" link with the "filter"
icon and the resolved concerns were already displayed...

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>