Bug 37354: Fix circ rule selections
Only expose the bookings calendar picker input if we have enough detail
already defined to pick the corresponding circulation rules.

We currently use the item pickup location for selecting which rules to
apply, this corresponds to hard coding CircControl=ItemHomeLibrary +
ItemHomeOrHolding=holdingbranch for this use case. We can't easily
change this to match those rules dynamically without rethinking the 'Any
item' functionality and displays in the flatpickr.  We may revisit this
at some point in the future should that become a requirement.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>