Bug 38221: Add Bootstrap styling to pagination in authority plugin search results
This patch updates the markup of pagination links in the authority
search popup so that they have Bootstrap styling consistent with other
pages (e.g. catalog search results).

The pagination markup has also been moved to a BLOCK so that it can be
reused instead of duplicated on the page.

To test, apply the patch and go to Cataloging -> New record (using the
basic MARC editor).

- Switch to tab 1 and click the tag editor link next to tag 100 subfield
  a.
- Perform a search in the popup window which will return multiple pages
  of results (an empty search will work with default testing data).
- Confirm that the pagination links are styled like catalog search
  results.
- Confirm that the search links work correctly and that the correct
  number is highlighted as you page through results.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>