Bug 35570: (QA follow-up) Remove unused variable definitions
 FAIL	Koha/ILL/Backend/Standard.pm
   FAIL	  critic

"$marc_isbn" is declared but not used at line 1179, column 5. Unused variables clutter code and make it harder to read.
"$marc_author" is declared but not used at line 1180, column 5. Unused variables clutter code and make it harder to read.
"$marc_title" is declared but not used at line 1181, column 5. Unused variables clutter code and make it harder to read.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>