Bug 38346: Make sidebar checkboxes consistent
This patch makes some changes to markup and CSS in order to make
checkboxes in sidebars more consistent with each other and better
styled.

To test apply the patch and rebuild the staff interface CSS. Check the
sidebar forms on these pages and verify that the checkbox inputs look
correct. The label should be on the same line as the checkbox.

- Serials -> Search serials -> Results: The "Has routing list" line.
- Administration -> Funds: The "Show my funds only" checkbox.
- Circulation -> Overdues: "Show any items currently checked out."
- Suggestions: "Include archived."

Sponsored-by: Athens County Public Libraries
Signed-off-by: Roman Dolny <roman.dolny@jezuici.pl>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>