Bug 38349: Fix style of sidebar form submit button on tags review page
This patch fixes the markup of the sidebar form on the tags
review page and removes an incorrect "btn-sm" class from the
submit button.

To test, apply the patch and go to Tools -> Tags.

In the sidebar form, confirm that the submit button has the same yellow
"primary" style as other submit buttons, and that it fills the width of
the sidebar.

Sponsored-by: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit f6c4663d85da8a1fc66d90051acf25fdcf5282d4)
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>