Bug 36640: Remove autoWidth: false for responsive tables
From Owen:
"""
Table columns don't collapse as you change the browser width.
They start collapsed if you narrow the window and reload the page. But they don't expand as you widen the browser window.
"""

I have not found a reason for this in the changelog of the responsive
plugin. Maybe there is another way to fix this?

Signed-off-by: Jake Deery <jake.deery@ptfs-europe.com>
Signed-off-by: Lisette Scheer <lisette@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>