Bug 38391: Redraw the filters row on table redraw
We used to adjust the filters row when the column visibility changed.
It was working, but at the initialization of the table this event if
fired once per columns. On the items table it means 36x.
When a filtering/sorting, the draw event is triggered once, but the
column-visibility.dt is called 36x

Test plan:
Try with and without this patch and compare the delay between the query
is received and the page is redrawn.

You can test with 1000 libraries on the items table.
"Show filters" then filters using the global search.
On Firefox you will notice that the response is received but the table
not redrawn directly.
With this patch it should be redrawn pretty fase.

Another test would be to add the following 2 lines:
+        let i = 0;
         table_dt.on("draw", function(){
+            console.log("redraw filters %s".format(i));
             if ( add_filters ) {
                 _dt_add_filters(this, table_dt, filters_options);
             }

Open the console and notice that "redraw filters" is logged only once.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>